refactor/이미지-업로드-수정-리팩터링 이미지 업로드 UI 헬퍼 모듈 분리
